# src/swagger.yaml
openapi: 3.0.3
info:
  title: NetScopeNMS REST API
  description: |
    AI 기반 네트워크 관리 시스템 (Network Management System) API 문서
    
    ## 주요 기능
    - **장비 관리**: 네트워크 장비 CRUD, SNMP 연결 테스트, 인터페이스 검색
    - **메트릭 수집**: 실시간 및 이력 데이터 조회 (CPU, 메모리, 트래픽 등)
    - **알람 관리**: 임계값 기반 알람 생성 및 관리
    - **AI 분석**: OpenAI 기반 근본 원인 분석, 예측, 리포트 생성
    - **사용자 관리**: JWT 기반 인증, 역할 기반 권한 관리
    
    ## 인증
    Bearer Token (JWT) 방식을 사용합니다.
    ```
    Authorization: Bearer <token>
    ```
    
    ## 오류 응답 형식
    ```json
    {
      "success": false,
      "error": "Error type",
      "message": "상세 메시지",
      "details": [] // 선택적 상세 정보
    }
    ```
    
  version: 1.0.0
  contact:
    name: NetScopeNMS Team
    email: support@netscopenms.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.netscopenms.com/api/v1
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /users/register:
    post:
      tags:
        - Authentication
      summary: 회원가입
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
                - password
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 50
                  pattern: ^[a-zA-Z0-9_]+$
                  example: john_doe
                  description: 사용자명 (영문, 숫자, 밑줄만 허용)
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 8
                  example: SecurePass123
                  description: 대문자, 소문자, 숫자 포함 필수
                role:
                  type: string
                  enum: [admin, user, viewer]
                  default: viewer
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/login:
    post:
      tags:
        - Authentication
      summary: 로그인
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: SecurePass123
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/me:
    get:
      tags:
        - Profile
      summary: 현재 사용자 프로필 조회
      responses:
        '200':
          description: 프로필 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
    put:
      tags:
        - Profile
      summary: 프로필 수정
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: 수정 성공
  /users/me/password:
    put:
      tags:
        - Profile
      summary: 비밀번호 변경
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 변경 성공
        '400':
          description: 현재 비밀번호 불일치

  # ==================== 사용자 관리 ====================
  /users:
    get:
      tags:
        - User Management
      summary: 사용자 목록 조회 (관리자)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, user, viewer]
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 사용자 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
    post:
      tags:
        - User Management
      summary: 사용자 생성 (관리자)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: 생성 성공
  /users/{id}:
    get:
      tags:
        - User Management
      summary: 사용자 상세 조회 (관리자)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 사용자 정보
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - User Management
      summary: 사용자 수정 (관리자)
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 수정 성공
    delete:
      tags:
        - User Management
      summary: 사용자 삭제 (관리자)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 삭제 성공
  /users/{id}/toggle:
    patch:
      tags:
        - User Management
      summary: 사용자 활성화/비활성화 (관리자)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: 토글 성공
  /users/{id}/reset-password:
    post:
      tags:
        - User Management
      summary: 비밀번호 초기화 (관리자)
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - newPassword
              properties:
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: 초기화 성공

  # ==================== 장비 관리 ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT 토큰을 사용한 인증
        
        로그인 후 받은 토큰을 `Authorization: Bearer <token>` 형식으로 전송
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: 페이지 번호
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: 페이지당 항목 수
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: 사용자 ID
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: 전체 항목 수
        page:
          type: integer
          description: 현재 페이지
        limit:
          type: integer
          description: 페이지당 항목 수
        totalPages:
          type: integer
          description: 전체 페이지 수
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, viewer]
        isActive:
          type: boolean
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, user, viewer]
          default: viewer
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, viewer]
        is_active:
          type: boolean
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            token:
              type: string
              description: JWT 토큰
...
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Validation failed
            message: 입력값 검증에 실패했습니다
            details:
              - field: email
                message: 이메일 형식이 올바르지 않습니다
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Unauthorized
            message: 인증이 필요합니다
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Not Found
            message: 요청한 리소스를 찾을 수 없습니다
    Conflict:
      description: 충돌 (중복 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Conflict
            message: 이미 존재하는 리소스입니다
...